Backend/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
faiss-cpu
numpy
together
openai
pydantic